[FIX] moved editcomponents subcomponents to subfolder
