+++ skill-db/frontend/src/components/EditComponents/UploadImageButton.vue
